work on gbuffer

+++ src/main/glsl/GBufferPipeline/main.vert
@@ -5,7 +5,7 @@
 layout(location = 0) in vec3 position;
 layout(location = 1) in vec3 normals;
 layout(location = 2) in vec2 textureCoordinates;
-layout(location = 3) in uint materialIndex;
+layout(location = 3) in float materialIndex;
 
 layout(location = 0) out vec3 readyPosition;
 layout(location = 1) out vec3 readyNormals;
@@ -37,6 +37,6 @@ void main(){
     readyPosition = (worldTransformData.worldMatrix*vec4(position, 1.0)).rgb;
     readyNormals = (vec4(normals, 0.0f)*worldTransformData.worldMatrix).xyz;
     readyTextureCoordinates = textureCoordinates;
-    rMaterialIndex = materialIndex;
+    rMaterialIndex = uint(materialIndex);
     gl_Position = fixVectorPositioning(worldTransformData.viewMatrix*worldTransformData.worldMatrix*vec4(position, 1.0));
 }
